Put useBackwardsCompatibleTheme in its own file
