+++ scripts/snapshots/es6Files.txt
@@ -221,6 +221,7 @@
   "es6/theme/graphicalItemIdentity.js",
   "es6/theme/legacyTheme.js",
   "es6/theme/lightTheme.js",
+  "es6/theme/useBackwardsCompatibleTheme.js",
   "es6/types.js",
   "es6/util",
   "es6/util/ActiveShapeUtils.js",
