+++ scripts/snapshots/typesFiles.txt
@@ -221,6 +221,7 @@
   "types/theme/graphicalItemIdentity.d.ts",
   "types/theme/legacyTheme.d.ts",
   "types/theme/lightTheme.d.ts",
+  "types/theme/useBackwardsCompatibleTheme.d.ts",
   "types/types.d.ts",
   "types/util",
   "types/util/ActiveShapeUtils.d.ts",
